-reorganized schema / resolvers
