+++ apollo-server/resolvers.js
@@ -1,5 +1,31 @@
-const db = require('../postgresDb/index');
+const UUID = require('graphql-type-uuid');
 
-const resolvers = {};
+const {
+  createUser,
+  authenticateUser,
+  getUser,
+  updateUser,
+  deleteUser,
+  Users,
+} = require('../postgresDb/auth/authHandlers');
 
+const resolvers = {
+  UUID,
+  Query: {
+    user: getUser,
+    authUser: authenticateUser,
+    users: Users,
+  },
+  Mutation: {
+    createUser,
+    updateUser,
+    deleteUser,
+  },
+  MutationResponse: {
+    __resolveType() {
+      return null;
+    },
+  },
+};
 
+module.exports = resolvers;
